requirements.txt
PyOpenSSL
gspread
google-api-python-client
oauth2client
auth
apiclient
xlsxwriter
konlpy
nltk
